separate transactions table into separate component
