+++ app/Http/Livewire/WalletVoterTable.php
@@ -8,6 +8,7 @@ use App\Http\Livewire\Concerns\HasTablePagination;
 use App\Models\Scopes\OrderByBalanceScope;
 use App\Models\Wallet;
 use App\ViewModels\ViewModelFactory;
+use App\ViewModels\WalletViewModel;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\View\View;
 use Livewire\Component;
@@ -21,15 +22,12 @@ final class WalletVoterTable extends Component
 
     public string $publicKey;
 
-    public string $username;
-
     /** @var mixed */
     protected $listeners = ['currencyChanged' => '$refresh'];
 
-    public function mount(string $publicKey, string $username): void
+    public function mount(WalletViewModel $wallet): void
     {
-        $this->publicKey = $publicKey;
-        $this->username  = $username;
+        $this->publicKey = $wallet->publicKey();
     }
 
     public function render(): View
